requirements.txt
# ── Framework ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

# ── Validation ────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-multipart==0.0.12

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Env ───────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Logging ───────────────────────────────────────────────
loguru==0.7.2

# ── Caching ───────────────────────────────────────────────
fastapi-cache2[redis]==0.2.2
redis==4.6.0

# ── Task Queue ────────────────────────────────────────────
celery==5.4.0

# ── HTTP Client ───────────────────────────────────────────
httpx==0.27.2

# ── AWS / Storage ─────────────────────────────────────────
boto3==1.35.36
Pillow==10.4.0

# ── Payments ──────────────────────────────────────────────
stripe==10.12.0


# ── Code Quality ──────────────────────────────────────────
ruff==0.6.9
mypy==1.11.2

# ── Security ───────────────────────────────────────────────
bandit==1.7.5
safety==3.2.0
pre-commit==4.0.1
